fix(ffi): align EmailAccountDto port types with Dart SSE codec (i32 not u16)
Dart's FRB SSE codec encodes int fields as i32 (4 bytes). The Rust
EmailAccountDto had smtp_port/imap_port as u16 (2 bytes), causing an
UnexpectedEof panic on deserialization whenever emailAccountSave was called.

- Change EmailAccountDto.smtp_port / imap_port: u16 → i32 in ffi.rs
- Add as u16 / as i32 casts at the EmailAccount boundary
- Update frb_generated.rs: SseDecode and SseEncode now use i32
- Add integration_test/email_account_test.dart: 8 round-trip tests
- Add integration_test dependency to mobileclaw_app pubspec.yaml

Result: all 8 Android integration tests pass on emulator-5554 (x86_64, API 36)

+++ mobileclaw-core/src/ffi.rs
@@ -74,9 +74,9 @@ pub struct SearchResultDto {
 pub struct EmailAccountDto {
     pub id: String,
     pub smtp_host: String,
-    pub smtp_port: u16,
+    pub smtp_port: i32,  // i32 matches Dart's int encoding via FRB SSE
     pub imap_host: String,
-    pub imap_port: u16,
+    pub imap_port: i32,  // i32 matches Dart's int encoding via FRB SSE
     pub username: String,
 }
 
@@ -294,9 +294,9 @@ impl AgentSession {
         let acc = EmailAccount {
             id: dto.id,
             smtp_host: dto.smtp_host,
-            smtp_port: dto.smtp_port,
+            smtp_port: dto.smtp_port as u16,
             imap_host: dto.imap_host,
-            imap_port: dto.imap_port,
+            imap_port: dto.imap_port as u16,
             username: dto.username,
         };
         self.secrets.put_email_account(&acc, &password).await.map_err(anyhow::Error::from)
@@ -315,9 +315,9 @@ impl AgentSession {
         Ok(Some(EmailAccountDto {
             id: acc.id,
             smtp_host: acc.smtp_host,
-            smtp_port: acc.smtp_port,
+            smtp_port: acc.smtp_port as i32,
             imap_host: acc.imap_host,
-            imap_port: acc.imap_port,
+            imap_port: acc.imap_port as i32,
             username: acc.username,
         }))
     }
